refactor(bichon): move api base url resolution into the service
rescan needs the same per-host / global / domain-derived resolution
reconcile-folders uses; hoist it to services::bichon so both commands
share one implementation.
