+++ src/commands/bichon/reconcile.rs
@@ -1,7 +1,8 @@
 use crate::config::Config;
-use crate::hosts::{Host, HostManager};
+use crate::hosts::HostManager;
 use crate::output::OutputFormat;
 use crate::services::bichon::api::{Account, BichonApiClient};
+use crate::services::bichon::derive_base_url;
 use crate::services::bichon::folder_filter::is_excluded;
 use eyre::{Result, WrapErr};
 use serde::Serialize;
@@ -55,7 +56,7 @@ pub async fn compute_reconcile(
         .get_resolved("bichon_api_token")?
         .filter(|v| !v.trim().is_empty())
         .ok_or_else(|| eyre::eyre!("bichon_api_token not set in config.toml"))?;
-    let base_url = derive_bichon_base_url(&config, &host_record)?;
+    let base_url = derive_base_url(&config, &host_record)?;
 
     let client = BichonApiClient::new(base_url, token)?;
     let mut accounts = client.list_accounts().await?;
@@ -128,33 +129,6 @@ pub async fn compute_reconcile(
             changed_accounts,
         },
     })
-}
-
-fn derive_bichon_base_url(config: &Config, host: &Host) -> Result<String> {
-    if let Some(per_host) = config.bichon_host_base_url(&host.name) {
-        return Ok(per_host);
-    }
-    if let Some(base_url) = config
-        .get("bichon_base_url")
-        .map(|v| v.trim().to_string())
-        .filter(|v| !v.is_empty())
-    {
-        return Ok(base_url.trim_end_matches('/').to_string());
-    }
-    let domain = config.domain();
-    if domain.is_empty() {
-        eyre::bail!(
-            "no Bichon base URL configured for host '{}'. Set [bichon.hosts.\"{}\"] base_url, or set bichon_base_url, or set domain in config.toml",
-            host.name,
-            host.name
-        )
-    }
-    let subdomain = config
-        .get("bichon_subdomain")
-        .map(|v| v.trim().to_string())
-        .filter(|v| !v.is_empty())
-        .unwrap_or_else(|| "bichon".to_string());
-    Ok(format!("https://{subdomain}.{domain}"))
 }
 
 fn emit_output(result: &ReconcileOutput, output: OutputFormat) -> Result<()> {
